+++ replacer-backend/src/test/java/es/bvalero/replacer/finder/immutable/QuotesAngularFinderTest.java
@@ -12,15 +12,16 @@ class QuotesAngularFinderTest {
 
     @Test
     void testRegexQuotesAngular() {
-        String quotes1 = "«yáy»";
-        String quotes2 = "«z, zz»";
-        String quotes3 = "«z\nz»";
-        String text = String.format("%s %s %s.", quotes1, quotes2, quotes3);
+        String quotes1 = "«xxx»";
+        String quotes2 = "«yyy\n"; // Truncated with new line
+        String quotes3 = "«zzz»";
+        String quotes4 = "«aaa"; // Truncated with end
+        String text = String.format("%s %s %s %s.", quotes1, quotes2, quotes3, quotes4);
 
         ImmutableFinder quotesFinder = new QuotesAngularFinder();
         List<Immutable> matches = quotesFinder.findList(text);
 
-        Set<String> expected = new HashSet<>(Arrays.asList(quotes1, quotes2));
+        Set<String> expected = new HashSet<>(Arrays.asList(quotes1, quotes3));
         Set<String> actual = matches.stream().map(Immutable::getText).collect(Collectors.toSet());
         Assertions.assertEquals(expected, actual);
     }